[GXA-339] Add an HCA experiments collection to the featured experiments (#553)
* Add an HCA experiments collection to the featured experiments

Also update the UI components with new versions and remove unused SVG files

* Update the project version to 17.2.0

Bump the version from 17.1.1 to 17.2.0 in the build configuration. This prepares the project for the next release with potential new changes or features.

* Update HCA image link and add new image resource

Updated the image URL and link text for the Human Cell Atlas summary card to reflect the new Data Portal designation. Added the corresponding new image resource to support this change.