[Chore] GXA-339 Reorder featured experiment collections (#554)
* Add an HCA experiments collection to the featured experiments

Also update the UI components with new versions and remove unused SVG files

* Update the project version to 17.2.0

Bump the version from 17.1.1 to 17.2.0 in the build configuration. This prepares the project for the next release with potential new changes or features.

* Update HCA image link and add new image resource

Updated the image URL and link text for the Human Cell Atlas summary card to reflect the new Data Portal designation. Added the corresponding new image resource to support this change.

* Reorder featured experiments cards initialization.

Adjusted the order of cards in the `featuredExperimentsCards` method to prioritize the "Human Cell Atlas" entry at the top. This ensures a consistent and intended display sequence for featured experiments.